Force credentials to be discoverable (i.e. stored on authenticator)
This is necessary for usernameless flows because if the user wants to authenticate, the server cannot return candiate credentials without first knowing who the user is. Therefore the user has to select a credential that is stored on one of their authenticators and present that to the server for verification.